\name{coRegnet-package}
\alias{coRegnet-package}
\alias{coRegnet}
\docType{package}
\title{
coRegnet : inference and interrogation co-regulation networks
}
\description{
coRegnet can be used to infer and analyse regulatory networks. Although it can be used with any network, it has it's own regulatory network inference routine based on a hybrid version of LICORN which combines both a discrete and a statistical model to infer regulatory network with an emphasis on regulator cooperativity.


}


\author{
Remy Nicolle <remy.c.nicolle AT gmail.com>
}
\references{
Chebil I, Nicolle R, Santini G, Rouveirol C & Elati M (2014) Hybrid Method Inference for the Construction of Cooperative Regulatory Network in Human. IEEE Trans Nanobioscience

Nicolle R, Elati M and Radvanyi F (2012) Network Transformation of Gene Expression for Feature Extraction. In pp 108-113.

Elati M, Neuvial P, Bolotin-Fukuhara M, Barillot E, Radvanyi F and Rouveirol C (2007) LICORN: learning cooperative regulation networks from gene expression data. Bioinformatics 23: 2407-2414
}
\keyword{ Biological network }
